+++ app/javascript/components/article/WhitePaper.jsx
@@ -1,27 +1,13 @@
 import React from 'react';
-import PropTypes from 'prop-types'
+import PropTypes from 'prop-types';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { withStyles } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
-import CardActions from '@material-ui/core/CardActions';
-import CardContent from '@material-ui/core/CardContent';
 import classNames from 'classnames';
 import Grid from '@material-ui/core/Grid';
-import CardHeader from '@material-ui/core/CardHeader';
-import MoreVertIcon from '@material-ui/icons/MoreVert';
-import IconButton from '@material-ui/core/IconButton';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-import Typography from '@material-ui/core/Typography';
-import TextField from '@material-ui/core/TextField';
-import Dropzone from 'react-dropzone'
-import Button from '@material-ui/core/Button';
 import Fab from '@material-ui/core/Fab';
 import AddIcon from '@material-ui/icons/Add';
-import DeleteIcon from '@material-ui/icons/Delete';
-import PhotoCamera from '@material-ui/icons/PhotoCamera';
-import CreateIcon from '@material-ui/icons/Create';
-import VisibilityIcon from '@material-ui/icons/Visibility';
+
+import Cell from './Cell'
 
 const styles = theme => ({
   root: {
@@ -48,141 +34,6 @@ const styles = theme => ({
     margin: theme.spacing.unit,
   },
 })
-
-class ImageDropZone extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      images: []
-    };
-
-    this.onDrop = this.onDrop.bind(this);
-    this.onCancel = this.onCancel.bind(this);
-  }
-
-  onDrop(images) {
-    this.setState({images});
-  }
-
-  onCancel() {
-    this.setState({
-      images: []
-    });
-  }
-
-  render() {
-    const images = this.state.images.map(image => (
-      <li key={image.name}>
-        {image.name} - {image.size} bytes
-      </li>
-    ))
-
-    return (
-      <div>
-        <h1>React S3 Image Uploader Sample</h1>
-        <Dropzone name={this.props.name} onDrop={this.onDrop} onFileDialogCancel={this.onCancel} accept="image/*">
-           {({getRootProps, getInputProps}) => (
-            <div {...getRootProps()}>
-              <input {...getInputProps()} />
-                <p>Drop files here, or click to select files</p>
-            </div>
-          )}
-        </Dropzone>
-        <aside>
-          <h4>Files</h4>
-          <ul>{images}</ul>
-        </aside>
-      </div>
-    );
-  }
-}
-
-class TabContainer extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  renderTextOrImage(type, id) {
-    if (type === 'text') {
-      return <TextField
-        name="article[bodys][]"
-        multiline
-        rowsMax="10"
-        margin="normal"
-      />
-    } else {
-      return <ImageDropZone name="article[bodys][]" />
-    }
-  }
-
-  render() {
-    return (
-      <React.Fragment>
-        {this.renderTextOrImage(this.props.type, this.props['data-key'])}
-      </React.Fragment>
-    )
-  }
-}
-
-TabContainer.propTypes = {
-  type: PropTypes.string.isRequired,
-};
-
-class Cell extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { value: 'text' };
-    this.handleChange = this.handleChange.bind(this);
-    this.onClick = this.onClick.bind(this);
-  }
-
-  handleChange(event, value) {
-    this.setState({ value });
-  };
-
-  onClick(id) {
-    this.props.deleteCell(id);
-  };
-
-  render() {
-    const { classes } = this.props;
-    const { value } = this.state;
-
-    return (
-      <Grid item xs={6} container alignItems="center" justify="center">
-        <Tabs value={value} onChange={this.handleChange}>
-          <Tab value='text' label="文章に入力する" />
-          <Tab value='image' label="画像をアップロードする" />
-        </Tabs>
-        <Card className={classes.card}>
-          <CardContent>
-            <TabContainer type={value} {...this.props}/>
-          </CardContent>
-          <CardActions>
-            <IconButton>
-              <CreateIcon />
-            </IconButton>
-            <label htmlFor="icon-button-file">
-              <IconButton component="span">
-                <PhotoCamera />
-              </IconButton>
-            </label>
-            <IconButton>
-              <VisibilityIcon />
-            </IconButton>
-            <IconButton aria-label="Delete">
-              <DeleteIcon size="small" onClick={() => this.onClick(this.props['data-key'])} />
-            </IconButton>
-          </CardActions>
-        </Card>
-      </Grid>
-    )
-  }
-}
-
-Cell.propTypes = {
-  deleteCell: PropTypes.func.isRequired,
-}
 
 class CellList extends React.Component {
   constructor(props) {
